+++ frontend/src/App.js
@@ -2,10 +2,11 @@ import { Routes, Route } from "react-router-dom"
 import Home from "./page/Home"
 import Login from "./page/Login"
 import Register from "./page/Register"
-import Feeding from "./page/Feeding"
 import { ProfilePage } from "./page/Profile";
 
 import PoemPage from "./page/PoemPage"
+import FeedingPage from "./page/FeedingPage"
+
 import { ProtectedLayout } from "./components/ProtectedLayout";
 import { HomeLayout } from "./components/HomeLayout";
 import  BasicSelect  from "./components/SelectChild";
@@ -23,7 +24,7 @@ function App() {
             <Route path="/dashboard" element={<ProtectedLayout />}>
                 {/*<Route path="/" element={<Home />}></Route>*/}
                 <Route path="profile" element={<ProfilePage />} />
-                <Route path="feeding" element={<Feeding />}></Route>
+                <Route path="feeding" element={<FeedingPage />}></Route>
                 <Route path="poemList" element={<PoemPage />}></Route>
                 <Route path="test" element={<BasicSelect />} />
             </Route>
